Add support for multiple formula conversion providers
The default is to use MathJax for .html flavour files and to use the
(unfortunately deprecated) Google charts API for other types, such as
.rss and .atom.  This can be overridden by configuration in the META
section of an article.  A "formulas" configuration overrides all
providers; a "formulas.foo" configuration overrides just the provider
for flavour 'foo'.  Valid providers are:

    local-tex       The old system, stores generated images on
                    pic.blog.plover.com

    mathjax         Uses mathjax

    72pines         Unimplemented support for inlined image from LaTeX server
                    at tex.72pines.org

    mathtran        Unimplemented support for inlined image from
                    www.mathtran.org

    google-charts   Inlined image from chart.googleapis.com

Still needs some work for proper handling of <formimg> elements.  In
particular, a lot of code in formula() should be moved into imglink().